refactor autograde helpers into separate helper
